requirements.txt
playwright>=1.40.0
keyring>=24.0.0
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
weasyprint>=60.0     # optional — for tailored PDF rendering
PyMuPDF>=1.24.0     # optional — for parsing base resume PDF to JSON
